Before End Of Day
Saving Changes.
Need to implement pathfinding.
Finished writing collision beginning for items on playfield.
Need to create a change in behaviour for enemy.
Have it so the enemy starts to seek as soon as the collecter drone, gets about 3 items.
Find a way to delete Items, and remove that collision.