Update
Fixed the error with deleting coin after falling in the corner pockets.

-Things to add
Fixing striker turning red, when in game phase
If Striker goes into holes, the positions of the coins are reset, and the turn is advanced.
Add UI for Score and Player Turn.
Make The positions of the coins invert to display the other player's turn---possible...